refactor: fix correlation IDs to be positive always refactor: move util methods to utils.go
