+++ internal/stagec1.go
@@ -2,7 +2,6 @@ package internal
 
 import (
 	"fmt"
-	"math"
 
 	"github.com/codecrafters-io/kafka-tester/internal/kafka_executable"
 	"github.com/codecrafters-io/kafka-tester/protocol"
@@ -27,7 +26,7 @@ func testSequentialRequests(stageHarness *test_case_harness.TestCaseHarness) err
 
 	requestCount := random.RandomInt(2, 5)
 	for i := 0; i < requestCount; i++ {
-		correlationId := int32(random.RandomInt(-math.MaxInt32, math.MaxInt32))
+		correlationId := getRandomCorrelationId()
 		request := kafkaapi.ApiVersionsRequest{
 			Header: kafkaapi.RequestHeader{
 				ApiKey:        18,
@@ -49,8 +48,8 @@ func testSequentialRequests(stageHarness *test_case_harness.TestCaseHarness) err
 		if err != nil {
 			return err
 		}
-		logger.Debugf("Hexdump of sent \"ApiVersions\" request: \n%v\n", protocol.GetFormattedHexdump(message))
-		logger.Debugf("Hexdump of received \"ApiVersions\" response: \n%v\n", protocol.GetFormattedHexdump(response))
+		logger.Debugf("Hexdump of sent \"ApiVersions\" request: \n%v\n", GetFormattedHexdump(message))
+		logger.Debugf("Hexdump of received \"ApiVersions\" response: \n%v\n", GetFormattedHexdump(response))
 
 		responseHeader, responseBody, err := kafkaapi.DecodeApiVersionsHeaderAndResponse(response, 3, logger)
 		if err != nil {
