+++ internal/utils.go
@@ -1,0 +1,63 @@
+package internal
+
+import (
+	"fmt"
+	"math"
+	"strings"
+
+	"github.com/codecrafters-io/tester-utils/random"
+)
+
+func getInvalidAPIVersion() int {
+	apiVersion := 1
+	for apiVersion <= 3 && apiVersion >= 0 {
+		apiVersion = random.RandomInt(0, math.MaxInt16)
+	}
+	return random.RandomElementFromArray([]int{apiVersion, -apiVersion})
+}
+
+func getRandomCorrelationId() int32 {
+	return int32(random.RandomInt(0, math.MaxInt32-1))
+}
+
+func GetFormattedHexdump(data []byte) string {
+	// This is used for logs
+	// Contains headers + vertical & horizontal separators + offset
+	// We use a different format for the error logs
+	var formattedHexdump strings.Builder
+	var asciiChars strings.Builder
+
+	formattedHexdump.WriteString("Idx  | Hex                                             | ASCII\n")
+	formattedHexdump.WriteString("-----+-------------------------------------------------+-----------------\n")
+
+	for i, b := range data {
+		if i%16 == 0 && i != 0 {
+			formattedHexdump.WriteString("| " + asciiChars.String() + "\n")
+			asciiChars.Reset()
+		}
+		if i%16 == 0 {
+			formattedHexdump.WriteString(fmt.Sprintf("%04x | ", i))
+		}
+		formattedHexdump.WriteString(fmt.Sprintf("%02x ", b))
+
+		// Add ASCII representation
+		if b >= 32 && b <= 126 {
+			asciiChars.WriteByte(b)
+		} else {
+			asciiChars.WriteByte('.')
+		}
+	}
+
+	// Pad the last line if necessary
+	if len(data)%16 != 0 {
+		padding := 16 - (len(data) % 16)
+		for i := 0; i < padding; i++ {
+			formattedHexdump.WriteString("   ")
+		}
+	}
+
+	// Add the final ASCII representation
+	formattedHexdump.WriteString("| " + asciiChars.String())
+
+	return formattedHexdump.String()
+}
